added adc, dma, and system + sample gpio for wokwi
paste code directly into wokwi with a "gpio.h" file for gpio.h

+++ include/adc.h
@@ -6,7 +6,7 @@
 
 typedef struct adc_inst_t adc_t;
 
-adc_t *adc_init(uint32_t channel);
+adc_t *adc_init(uint8_t channel);
 
 uint16_t adc_read_value(adc_t *adc);
 uint16_t adc_read_avg(adc_t *adc, uint8_t samples);
